Moved all remaining clusters
