+++ tcc/inc/clusters/10A.h
@@ -1,0 +1,8 @@
+#ifndef TCC_10A_H
+#define TCC_10A_H
+
+void Clusters_Get10A();
+
+void Cluster_Write_10A();
+
+#endif
