+++ tcc/inc/clusters/10B.h
@@ -1,0 +1,8 @@
+#ifndef TCC_10B_H
+#define TCC_10B_H
+
+void Clusters_Get10B(int i, int j);
+
+void Cluster_Write_10B();
+
+#endif
