+++ tcc/inc/clusters/10K.h
@@ -1,0 +1,10 @@
+#ifndef TCC_10K_H
+#define TCC_10K_H
+
+void Clusters_Get10K();
+
+int is_particle_in_9K(int id_9k, int id_particle);
+
+void Cluster_Write_10K(int id_9K, int ep);
+
+#endif
