+++ tcc/inc/clusters/9K.h
@@ -1,0 +1,8 @@
+#ifndef TCC_9K_H
+#define TCC_9K_H
+
+void Clusters_Get9K();
+
+void Cluster_Write_9k(const int trial[]);
+
+#endif
