+++ tcc/src/main.c
@@ -5,7 +5,6 @@
 #include "globals.h"
 #include "bonds.h"
 #include "rings.h"
-#include "clusters.h"
 #include "output.h"
 #include "stats.h"
 #include "tools.h"
@@ -14,23 +13,27 @@
 #include "7K.h"
 #include "8A.h"
 #include "8B.h"
+#include "8K.h"
+#include "9A.h"
+#include "9B.h"
+#include "9K.h"
+#include "10A.h"
+#include "10K.h"
+#include "10W.h"
 #include "11A.h"
 #include "11B.h"
 #include "11C.h"
 #include "11W.h"
 #include "11F.h"
 #include "12A.h"
+#include "12B.h"
 #include "12E.h"
 #include "12K.h"
-#include "8K.h"
-#include "9A.h"
+#include "13B.h"
 #include "HCP.h"
 #include "FCC.h"
 #include "BCC15.h"
 #include "BCC9.h"
-#include "13B.h"
-#include "12B.h"
-#include "10W.h"
 
 int num_cluster_types = 39;
